// src/types.ts
export enum TYPES {
    FIELD_SERVICE = "FIELD_SERVICE",
    GAME_SERVICE = "GAME_SERVICE",
    PLAYER_SERVICE = "PLAYER_SERVICE",
    UNIT_SERVICE = "UNIT_SERVICE",

    FIELD_REPOSITORY = "FIELD_REPOSITORY",
    GAME_REPOSITORY = "GAME_REPOSITORY",
    PLAYER_REPOSITORY = "PLAYER_REPOSITORY",
    UNIT_REPOSITORY = "UNIT_REPOSITORY",
    ACTION_TYPE_REPOSITORY = "ACTIION_TYPE_REPOSITORY",

    FIELD_JSON_MAPPER = "FIELD_JSON_MAPPER",
    GAME_JSON_MAPPER = "GAME_JSON_MAPPER",
    PLAYER_JSON_MAPPER = "PLAYER_JSON_MAPPER",
    UNIT_JSON_MAPPER = "UNIT_JSON_MAPPER",

    FIELD_ALGORITHM_SERVICE = "MOVEMENT_SERVICE",
    ACTION_SERVICE = "ACTION_SERVICE"
};